Allow filtering negative ratings from ials
The implicit als model assumes that non-zeros passed in are positive ratings.
For datasets like movielens passing in 1 and 2 star reviews in the training set
causes these to be learned as positive preferences - which leads to poor results.

This adds a 'train_switch_positive' option to the ialswrapper, which when set
will remove ratings from the training set less than this value.

Applying this change to the warm-start example and the iALS model starts to perform
more competitively.